First Commit
Well there are some blogs and sites where you can find an example about
paging on a data grid or advanceddata gridd.

There are also a lot of blogs about filtering on fledata gridsds or
flex AdvancedDatagrid.

…... there are no examples about paging AND filtering on flex Datagrids
or Flex AdvancedDatagrids.

Now there is !

When I was working on a solution for paging a big amount of information
in a flex Project the people I was building it for asked me if it was
possible to keep the filtering functionality in.

At first I thought it was not possible. But.…. after a good evening at
trying, coding and testing at home I had sucsuccessd paging and
filtering on a Adobe Flex Datagrid (or AdvancedDatagrid) is possible.

There is only one big issue, on a big amount of row's it can become
slow.